fix: executeScript can't use class method
