+++ src/background/handlers/storage-handler.ts
@@ -1,12 +1,13 @@
 import { StorageData } from '../../shared/types';
 import { ExtensionError } from '../../shared/utils/error-handling';
+import { clearStorage, extractStorageData, injectStorageData } from '../services/storageData-service';
 
 export class StorageHandler {
   async getStorageData(tabId: number): Promise<StorageData> {
     try {
       const results = await chrome.scripting.executeScript({
         target: { tabId },
-        func: this.extractStorageData
+        func: extractStorageData
       });
 
       return results?.[0]?.result || { localStorage: {}, sessionStorage: {} };
@@ -20,7 +21,7 @@ export class StorageHandler {
     try {
       await chrome.scripting.executeScript({
         target: { tabId },
-        func: this.injectStorageData,
+        func: injectStorageData,
         args: [data.localStorage, data.sessionStorage]
       });
     } catch (error) {
@@ -32,81 +33,10 @@ export class StorageHandler {
     try {
       await chrome.scripting.executeScript({
         target: { tabId },
-        func: this.clearStorage
+        func: clearStorage
       });
     } catch (error) {
       throw new ExtensionError(`Failed to clear storage data: ${error}`);
-    }
-  }
-
-  private extractStorageData(): StorageData {
-    try {
-      const localStorageData: Record<string, string> = {};
-      const sessionStorageData: Record<string, string> = {};
-
-      // Extract localStorage
-      for (let i = 0; i < localStorage.length; i++) {
-        const key = localStorage.key(i);
-        if (key) {
-          const value = localStorage.getItem(key);
-          if (value !== null) {
-            localStorageData[key] = value;
-          }
-        }
-      }
-
-      // Extract sessionStorage
-      for (let i = 0; i < sessionStorage.length; i++) {
-        const key = sessionStorage.key(i);
-        if (key) {
-          const value = sessionStorage.getItem(key);
-          if (value !== null) {
-            sessionStorageData[key] = value;
-          }
-        }
-      }
-
-      return {
-        localStorage: localStorageData,
-        sessionStorage: sessionStorageData
-      };
-    } catch (error) {
-      console.error('Error extracting storage data:', error);
-      return { localStorage: {}, sessionStorage: {} };
-    }
-  }
-
-  private injectStorageData(localData: Record<string, string>, sessionData: Record<string, string>): boolean {
-    try {
-      // Clear existing storage
-      localStorage.clear();
-      sessionStorage.clear();
-
-      // Restore localStorage
-      Object.entries(localData).forEach(([key, value]) => {
-        localStorage.setItem(key, value);
-      });
-
-      // Restore sessionStorage
-      Object.entries(sessionData).forEach(([key, value]) => {
-        sessionStorage.setItem(key, value);
-      });
-
-      return true;
-    } catch (error) {
-      console.error('Error injecting storage data:', error);
-      return false;
-    }
-  }
-
-  private clearStorage(): boolean {
-    try {
-      localStorage.clear();
-      sessionStorage.clear();
-      return true;
-    } catch (error) {
-      console.error('Error clearing storage:', error);
-      return false;
     }
   }
 }
